consolidate log service request logic
